+++ pkg/utils/object/compare/suite_test.go
@@ -1,0 +1,13 @@
+package compare
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestSuite(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "utils/object/compare test suite")
+}